+++ docs/versionhistory.rst
@@ -15,6 +15,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Added the ability to set the token count of a ``CapacityLimiter`` to zero
   (`#1019 <https://github.com/agronholm/anyio/pull/1019>`_; requires Python 3.10 or
   later when using Trio)
+- Dropped ``sniffio`` as a direct dependency and added the ``get_available_backends()``
+  function (`#1021 <https://github.com/agronholm/anyio/pull/1021>`_)
 - Fixed ``Process.stdin.send()`` not raising ``ClosedResourceError`` and
   ``BrokenResourceError`` on asyncio. Previously, a non-AnyIO exception was raised in
   such cases (`#671 <https://github.com/agronholm/anyio/issues/671>`_; PR by